+++ src/rules/flake8_errmsg/rules.rs
@@ -3,7 +3,40 @@ use rustpython_ast::{Constant, Expr, ExprKind};
 use crate::ast::types::Range;
 use crate::checkers::ast::Checker;
 use crate::registry::{Diagnostic, Rule};
-use crate::violations;
+use crate::violation::Violation;
+
+use crate::define_violation;
+use ruff_macros::derive_message_formats;
+
+define_violation!(
+    pub struct RawStringInException;
+);
+impl Violation for RawStringInException {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("Exception must not use a string literal, assign to variable first")
+    }
+}
+
+define_violation!(
+    pub struct FStringInException;
+);
+impl Violation for FStringInException {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("Exception must not use an f-string literal, assign to variable first")
+    }
+}
+
+define_violation!(
+    pub struct DotFormatInException;
+);
+impl Violation for DotFormatInException {
+    #[derive_message_formats]
+    fn message(&self) -> String {
+        format!("Exception must not use a `.format()` string directly, assign to variable first")
+    }
+}
 
 /// EM101, EM102, EM103
 pub fn string_in_exception(checker: &mut Checker, exc: &Expr) {
@@ -18,7 +51,7 @@ pub fn string_in_exception(checker: &mut Checker, exc: &Expr) {
                     if checker.settings.rules.enabled(&Rule::RawStringInException) {
                         if string.len() > checker.settings.flake8_errmsg.max_string_length {
                             checker.diagnostics.push(Diagnostic::new(
-                                violations::RawStringInException,
+                                RawStringInException,
                                 Range::from_located(first),
                             ));
                         }
@@ -28,7 +61,7 @@ pub fn string_in_exception(checker: &mut Checker, exc: &Expr) {
                 ExprKind::JoinedStr { .. } => {
                     if checker.settings.rules.enabled(&Rule::FStringInException) {
                         checker.diagnostics.push(Diagnostic::new(
-                            violations::FStringInException,
+                            FStringInException,
                             Range::from_located(first),
                         ));
                     }
@@ -39,7 +72,7 @@ pub fn string_in_exception(checker: &mut Checker, exc: &Expr) {
                         if let ExprKind::Attribute { value, attr, .. } = &func.node {
                             if attr == "format" && matches!(value.node, ExprKind::Constant { .. }) {
                                 checker.diagnostics.push(Diagnostic::new(
-                                    violations::DotFormatInException,
+                                    DotFormatInException,
                                     Range::from_located(first),
                                 ));
                             }
