--- src/violations.rs
@@ -1,39 +1,0 @@
-#![allow(clippy::useless_format)]
-
-use ruff_macros::derive_message_formats;
-
-use crate::define_violation;
-
-use crate::violation::Violation;
-
-// flake8-errmsg
-
-define_violation!(
-    pub struct RawStringInException;
-);
-impl Violation for RawStringInException {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Exception must not use a string literal, assign to variable first")
-    }
-}
-
-define_violation!(
-    pub struct FStringInException;
-);
-impl Violation for FStringInException {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Exception must not use an f-string literal, assign to variable first")
-    }
-}
-
-define_violation!(
-    pub struct DotFormatInException;
-);
-impl Violation for DotFormatInException {
-    #[derive_message_formats]
-    fn message(&self) -> String {
-        format!("Exception must not use a `.format()` string directly, assign to variable first")
-    }
-}
